Hopefully close to the end
This is close to the ending product. Ui and design work flawlessly however adding user profiles is not currently available mostly for the reason that it wont be able to work in such a way because it needs to have a database. further testing needed to be finalized also comments need to be added where people have coded

Co-Authored-By: Philip Noulis <148959003+P-NOULIS@users.noreply.github.com>